extract github-actions.js
