+++ .gitignore
@@ -25,3 +25,5 @@ windows/**/build/
 windows/**/bld/
 windows/**/[Bb]in/
 windows/**/[Oo]bj/
+
+/action/node_modules/
